requirements.txt
# ==============================================================================
# AI Resume Screening System — Dependencies
# 100% Free & Open-Source. No paid APIs or cloud services.
# ==============================================================================

# --- API Framework ---
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20        # file upload support

# --- Data Validation ---
pydantic==2.10.3

# --- NLP ---
spacy==3.8.3
# Run after install: python -m spacy download en_core_web_sm

# --- Embeddings (local, free) ---
sentence-transformers==3.3.1    # downloads all-MiniLM-L6-v2 on first use

# --- Vector Search ---
faiss-cpu==1.9.0.post1

# --- ML Utilities ---
scikit-learn==1.6.0
numpy==1.26.4

# --- Document Parsing ---
PyMuPDF==1.25.1                 # PDF parsing (fitz)
python-docx==1.1.2              # DOCX parsing

# --- Frontend ---
streamlit==1.41.1
plotly==5.24.1                  # charts in Streamlit
requests==2.32.3                # UI → API HTTP calls
fpdf2==2.8.6                    # PDF report generation

# --- Database ---
# sqlite3 is part of Python stdlib — no install needed

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1                   # async test client for FastAPI

# --- Dev Utilities ---
python-dotenv==1.0.1
